feat: replace rest client in another file
