+++ src/app/[locale]/page.tsx
@@ -1,77 +1,12 @@
-'use client';
+import { redirect } from '@i18n/navigation';
+import { routesList } from '@data/routes-list';
 
-import { useState } from 'react';
-import { MethodSelector } from '@components';
-import { EndpointInput } from '@components';
-import { HeadersEditor } from '@components';
-import { BodyEditor } from '@components';
-import { ResponseViewer } from '@components';
+export default async function RootPage({
+  params,
+}: {
+  params: Promise<{ locale: string }>;
+}) {
+  const { locale } = await params;
 
-type APIResponse = { error: string } | { status: number; data: unknown } | null;
-
-export default function RestPage() {
-  const [method, setMethod] = useState<string>('GET');
-  const [endpoint, setEndpoint] = useState('');
-  const [headers, setHeaders] = useState<{ key: string; value: string }[]>([]);
-  const [body, setBody] = useState('');
-  const [response, setResponse] = useState<APIResponse>(null);
-  const [isLoading, setIsLoading] = useState(false);
-
-  const handleSubmit = async () => {
-    if (!endpoint.trim()) return;
-
-    setIsLoading(true);
-    try {
-      const res = await fetch(endpoint, {
-        method,
-        headers: headers.reduce(
-          (acc, h) => {
-            acc[h.key] = h.value;
-            return acc;
-          },
-          {} as Record<string, string>
-        ),
-        body: method === 'GET' || !body ? undefined : body,
-      });
-
-      const data = await res.json();
-      setResponse({ status: res.status, data });
-    } catch (err) {
-      setResponse({
-        error: err instanceof Error ? err.message : 'Unknown error',
-      });
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  return (
-    <div className="p-6 max-w-4xl mx-auto space-y-6">
-      <h1 className="text-2xl font-bold">REST Client</h1>
-
-      <div className="space-y-4">
-        <MethodSelector value={method} onChange={setMethod} />
-        <EndpointInput value={endpoint} onChange={setEndpoint} />
-        <HeadersEditor
-          headers={headers}
-          onAdd={(key, value) => setHeaders([...headers, { key, value }])}
-        />
-        <BodyEditor value={body} onChange={setBody} />
-        <button
-          onClick={handleSubmit}
-          disabled={isLoading || !endpoint}
-          className="px-4 py-2 bg-violet-500 text-violet-950 rounded hover:bg-violet-600 disabled:opacity-50"
-        >
-          {isLoading ? 'Sending...' : 'Send Request'}
-        </button>
-      </div>
-
-      {response && (
-        <div className="mt-8 p-4 border rounded bg-gray-50">
-          <h2 className="text-lg font-semibold mb-2">Response</h2>
-          <ResponseViewer data={response} />
-        </div>
-      )}
-    </div>
-  );
+  redirect({ href: `/${routesList.main}`, locale: locale });
 }
